Split Key & MouseButton classes in two files
